farmaa_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
slowapi==0.1.9
google-auth==2.29.0
requests==2.31.0
pydantic[email]==2.7.0
pytest==8.1.1
httpx==0.27.0
firebase-admin==6.5.0
